Experiment can be used in a with-statement
I noticed that sometimes my latest iterations aren't logged in sqlite.
It turns out it's due to the buffer of the writer not being properly closed.
Experiments.__del__ is supposed to take care of it. But this method
is called upon garbage collection of the Experiment object, which,
according to [1], isn't garanteed:

> An implementation is allowed to postpone garbage collection
> or omit it altogether

With this __exit__ pattern we can ensure that the data is deterministically
written.
I also keep the destructor for users who don't call it properly (or for
legacy codebases)

[1]: https://docs.python.org/2/reference/datamodel.html